# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Service Peminjaman API",
        "description": "Dokumentasi layanan peminjaman buku untuk Tugas 2 IAE",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "Current Server"
        }
    ],
    "paths": {
        "/api/v1/loans": {
            "get": {
                "tags": [
                    "Loans"
                ],
                "summary": "Ambil semua transaksi peminjaman",
                "operationId": "11067bb9b862fd7effd5a9394c6aa29e",
                "responses": {
                    "200": {
                        "description": "Data retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Loans"
                ],
                "summary": "Buat transaksi peminjaman baru",
                "operationId": "67746018747d878b72f133d469781040",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "member_id",
                                    "book_id"
                                ],
                                "properties": {
                                    "member_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "book_id": {
                                        "type": "integer",
                                        "example": 5
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Data created successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            }
        },
        "/api/v1/loans/{id}": {
            "get": {
                "tags": [
                    "Loans"
                ],
                "summary": "Ambil detail satu transaksi",
                "operationId": "93e76696ff0b06ec950dd75e35c89c1b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID Transaksi Peminjaman",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Data retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Resource not found",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            }
        },
        "/api/v1/loans/{id}/return": {
            "post": {
                "tags": [
                    "Loans"
                ],
                "summary": "Proses pengembalian buku",
                "operationId": "2e6a11e9ff54d0caf7c65d4173719a98",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID Transaksi Peminjaman",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Book returned successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Resource not found",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "ApiKeyAuth": {
                "type": "apiKey",
                "description": "API Key (NIM) untuk autentikasi service",
                "name": "X-IAE-KEY",
                "in": "header"
            }
        }
    },
    "tags": [
        {
            "name": "Loans",
            "description": "Loans"
        }
    ]
}